Renamed fosite utils to oauth2 utils
