+++ controllers/moderator/oauth.go
@@ -15,9 +15,9 @@ import (
 	"github.com/NyaaPantsu/nyaa/models/oauth_client"
 	"github.com/NyaaPantsu/nyaa/templates"
 	"github.com/NyaaPantsu/nyaa/utils/format"
-	"github.com/NyaaPantsu/nyaa/utils/fosite/manager"
 	"github.com/NyaaPantsu/nyaa/utils/log"
 	msg "github.com/NyaaPantsu/nyaa/utils/messages"
+	"github.com/NyaaPantsu/nyaa/utils/oauth2/manager"
 	"github.com/NyaaPantsu/nyaa/utils/validator"
 	"github.com/NyaaPantsu/nyaa/utils/validator/api"
 	"github.com/gin-gonic/gin"
