+++ utils/oauth2/manager/manager_sql.go
@@ -9,7 +9,7 @@ import (
 	"github.com/NyaaPantsu/nyaa/models"
 	"github.com/NyaaPantsu/nyaa/models/oauth_client"
 	"github.com/NyaaPantsu/nyaa/utils/format"
-	"github.com/NyaaPantsu/nyaa/utils/fosite/client"
+	"github.com/NyaaPantsu/nyaa/utils/oauth2/client"
 	"github.com/ory/fosite"
 	"github.com/pborman/uuid"
 	"github.com/pkg/errors"
